add NET/EET TIFF/Google TIFF to workflow

+++ n0r.csh
@@ -1,4 +1,5 @@
 # Special script for gridding the NEXRAD composites
+# set echo
 
 setenv STARTTIME `date -u +'%Y%m%d%H%M%S'`
 
@@ -56,7 +57,7 @@ if ($realtime == "t") then
 endif
 
 # Convert it to TIF for algorithm work
-convert -compress none $fp n0r_$$_in.tif
+convert -compress none $fp n0r_$$_in.tif >& /dev/null
 rm -f $fp
 
 # Now generate NET
@@ -83,40 +84,47 @@ EOF
 
 if ($realtime == "t") then
   # Insert NET file
-  convert -define PNG:preserve-colormap $fp test_$$.png
+  convert -define PNG:preserve-colormap $fp test_$$.png >& /dev/null
   pqinsert -p "gis cr ${ftime} gis/images/4326/USCOMP/net_ bogus png" test_$$.png
   rm -f test_$$.png
 endif
 
 # Convert it to TIF for algorithm work
-convert -compress none $fp net_$$_in.tif
+convert -compress none $fp net_$$_in.tif >& /dev/null
 rm -f $fp
 
 # Clean it! 
 python gdal_clean.py $$ $1$2$3$4$5
 
 # Convert result file to PNG and send it on its way!
-convert -define PNG:preserve-colormap n0r_$$_out.tif test_$$.png
+convert -define PNG:preserve-colormap n0r_$$_out.tif test_$$.png >& /dev/null
 pqinsert -p "gis ${routes} ${ftime} gis/images/4326/USCOMP/n0r_ GIS/uscomp/n0r_${ftime}.png png" test_$$.png
 
 # Also insert the world file, so that the archive gets it!
 python n0r_gentfw.py ${ftime} n0r
-pqinsert -i -p "gis ${routes} ${ftime} gis/images/4326/USCOMP/n0r_ GIS/uscomp/n0r_${ftime}.wld wld" n0r${ftime}.tfw
+pqinsert -p "gis ${routes} ${ftime} gis/images/4326/USCOMP/n0r_ GIS/uscomp/n0r_${ftime}.wld wld" n0r${ftime}.tfw
 
 if ($realtime == "t") then
   # Generate GeoTIFF
-  geotifcp -e n0r.tfw n0r_$$_out.tif n0r_$$_out.gtif
+  geotifcp -e n0r.tfw n0r_$$_out.tif n0r_$$_out.gtif >& /dev/null
+  geotifcp -e n0r.tfw net_$$_in.tif net_$$_in.gtif >& /dev/null
 
   # Generate google maps variant
   cp n0r.tfw n0r_$$_out.tfw
-  gdalwarp  -q -s_srs EPSG:4326 -t_srs '+proj=merc +a=6378137 +b=6378137 +lat_ts=0.0 +lon_0=0.0 +x_0=0.0 +y_0=0 +k=1.0 +units=m +nadgrids=@null +no_def' n0r_$$_out.tif google_n0r_$$_out.tif
+  gdalwarp  -q -s_srs EPSG:4326 -t_srs EPSG:3857 n0r_$$_out.tif google_n0r_$$_out.tif
   rm -f n0r_$$_out.tfw
+
+    gdalwarp  -q -s_srs EPSG:4326 -t_srs EPSG:3857 net_$$_in.gtif google_net_$$_in.tif
 
   gzip -c n0r_$$_out.tif > n0r_$$_out.tif.Z
   pqinsert -p "gis r ${ftime} gis/images/4326/USCOMP/n0r_ bogus tif.Z" n0r_$$_out.tif.Z
 
   gzip -c google_n0r_$$_out.tif > google_n0r_$$_out.tif.Z
   pqinsert -p "gis r ${ftime} gis/images/900913/USCOMP/n0r_ bogus tif.Z" google_n0r_$$_out.tif.Z
+
+  gzip -c google_net_$$_in.tif > google_net_$$_in.tif.Z
+  pqinsert -p "gis r ${ftime} gis/images/900913/USCOMP/net_ bogus tif.Z" google_net_$$_in.tif.Z
+  rm -f google_net_$$_in.tif.Z google_net_$$_in.tif
 
   gzip -c n0r_$$_out.gtif > n0r_$$_out.gtif.Z
   pqinsert -p "gis r ${ftime} gis/images/4326/USCOMP/n0r_ bogus gtif.Z" n0r_$$_out.gtif.Z
@@ -125,7 +133,9 @@ if ($realtime == "t") then
   cd tmp/
   foreach i (10 9 8 7 6 5 4 3 2 1 0)
     set j = `echo "${i} + 1" | bc `
-    mv n0r_${i}.gif n0r_${j}.gif
+    if ( -f n0r_${i}.gif ) then
+      mv n0r_${i}.gif n0r_${j}.gif
+    endif
   end
   convert ../test_$$.png n0r_0.gif
   gifsicle -U --transparent "#000" --loopcount=0 --delay=100 n0r_11.gif n0r_10.gif  n0r_9.gif n0r_8.gif n0r_7.gif n0r_6.gif n0r_5.gif n0r_4.gif n0r_3.gif n0r_2.gif n0r_1.gif n0r_0.gif  -o anim.gif >& /dev/null
@@ -138,6 +148,7 @@ if ($realtime == "t") then
   cd ../
 endif
 
+rm -f net_$$_in.gtif
 rm -f n0r${ftime}.tfw test_$$.png net_$$_in.tif n0r_$$_in.tif n0r_$$_out.tif n0r_$$_out.tif.Z n0r_$$_out.gtif.Z google_n0r_$$_out.tif.Z google_n0r_$$_out.tif n0r_$$_out.gtif
 
 python scripts/create_metadata.py US $1 $2 $3 $4 $5 N0R $STARTTIME
